Changed naming of the httpServer actor to better understandable one. Also setup unit testing
