+++ runtime-api/src/main/scala/io/coral/actors/CoralActorFactory.scala
@@ -27,7 +27,7 @@ object CoralActorFactory {
         case "zscore"     => ZscoreActor(json)
         case "stats"      => StatsActor(json)
         case "lookup"     => LookupActor(json)
-        case "httpserver" => HttpServerActor(json)
+        case "httpserver" => HttpBroadcastActor(json)
         case "httpclient" => HttpClientActor(json)
         case "cassandra"  => CassandraActor(json)
         case "threshold"  => ThresholdActor(json)
