+++ runtime-api/src/test/scala/io/coral/actors/transform/HttpBroadcastActorSpec.scala
@@ -9,7 +9,7 @@ import org.scalatest.{BeforeAndAfterAll, Matchers, WordSpecLike}
 /**
  * Created by us73ut on 3/25/15.
  */
-class HttpServerActorSpec(_system: ActorSystem)
+class HttpBroadcastActorSpec(_system: ActorSystem)
   extends TestKit(_system)
   with ImplicitSender
   with WordSpecLike
